Improve handling of days without time slots in availability
Disable selection and update styling for days with no time slots in the week calendar
Adjust week navigation to select the first day with available slots if the corresponding day in the new week has no slots.
Enhance "copy schedule to week" to apply only to future days and and fix existing bookings slot to be duplicate.
Prevent generation of time slots that are in the past.